requirements.txt
simpy
google-generativeai
python-dotenv
google-genai